requirements.txt
Flask==3.0.0
pyserial==3.5
requests==2.31.0

